Update ish_proper_unload.script
ipairs should be used with arrays, garuntees order and is faster.

not has a higher priority than and  original logic only and always evaluate to true if IsWeapon was false, making the fake_ammo check redundant. and letting mele weapons thru to the other functions. as a result that logic kept trying to unload knives spamming the log. This logic works as intended and is probably the easiest to read.